chore(cext): fix 4 clippy regressions from msgpack spike

+++ crates/rubyrs-cext/src/stubs/types.rs
@@ -158,9 +158,12 @@ pub unsafe extern "C" fn RHASH_SIZE(v: Value) -> c_long {
 // ===== msgpack-ruby additions =====
 
 /// long long -> Integer. rubyrs's Number is i64; identical to rb_ll2num.
+/// Rust's `std::ffi::c_longlong` is defined as `i64` (per the Rust
+/// ABI types — the C standard only mandates ≥64 bits, but the Rust
+/// alias picks the exact-i64 contract), so no cast is needed.
 #[unsafe(no_mangle)]
 pub unsafe extern "C" fn rb_ll2inum(n: c_longlong) -> Value {
-    with_state(|st| st.intern(CValue::Int(n as i64)))
+    with_state(|st| st.intern(CValue::Int(n)))
 }
 
 /// unsigned long long -> Integer. Truncates to i64 (rubyrs has no
